requirements.txt
et-xmlfile==1.1.0
MouseInfo==0.1.3
numpy==1.21.1
opencv-python==4.5.3.56
openpyxl==3.0.7
Pillow==8.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PyTweening==1.0.3
